requirements.txt
aiohttp
aiofiles
aioshutil
anytree
apscheduler
asyncio
beautifulsoup4
cloudscraper
dnspython
feedparser
flask
gevent
gunicorn
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
httpx
lxml
motor
natsort
pillow
psutil
pymongo
kurigram
python-dotenv
python-magic
requests
tenacity
tgcrypto
urllib3
uvloop
xattr
requests
